cloud_function/requirements.txt
google-cloud-storage
google-genai
requests
